+++ common/src/main/java/me/pandamods/fallingtrees/mixin/BlockBehaviourMixin.java
@@ -1,5 +1,7 @@
 package me.pandamods.fallingtrees.mixin;
 
+import me.pandamods.fallingtrees.api.TreeHandler;
+import me.pandamods.fallingtrees.config.FallingTreesConfig;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
@@ -10,10 +12,15 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import java.util.Optional;
+
 @Mixin(BlockBehaviour.class)
 public class BlockBehaviourMixin {
 	@Inject(method = "getDestroyProgress", at = @At("TAIL"), cancellable = true)
 	private void getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos, CallbackInfoReturnable<Float> cir) {
-		// Todo: add dynamic breaking speed 
+		if (FallingTreesConfig.getCommonConfig().dynamicMiningSpeed.disable) return;
+		if (player == null || !TreeHandler.canPlayerChopTree(player)) return;
+		Optional<Float> miningSpeedOpt = TreeHandler.getMiningSpeed(player, pos, cir.getReturnValue());
+		miningSpeedOpt.ifPresent(cir::setReturnValue);
 	}
 }